feat(ui): expose bucketid (#18319)
* refactor: extract bucket card actions and meta into separate components

* feat: allow bucket ID to be copied

* chore: cleanup

* chore: update changelog

* feat: add copiable ID to demo data bucket cards
