+++ ui/src/buckets/components/BucketCardMeta.scss
@@ -1,0 +1,20 @@
+.copy-bucket-id {
+  transition: color 0.25s ease;
+
+  &:hover {
+    cursor: pointer;
+    color: $c-pool;
+  }
+}
+
+.copy-bucket-id--helper {
+  color: $g13-mist;
+  transition: opacity 0.25s ease;
+  opacity: 0;
+  display: inline-block;
+  margin-left: $cf-marg-b;
+}
+
+.copy-bucket-id:hover .copy-bucket-id--helper {
+  opacity: 1;
+}
