+++ ui/src/buckets/components/DemoDataBucketCard.tsx
@@ -1,7 +1,14 @@
 // Libraries
 import React, {FC} from 'react'
 import {withRouter, WithRouterProps} from 'react-router'
+import CopyToClipboard from 'react-copy-to-clipboard'
 import {connect} from 'react-redux'
+
+// Constants
+import {
+  copyToClipboardSuccess,
+  copyToClipboardFailed,
+} from 'src/shared/copy/notifications'
 
 // Components
 import {
@@ -20,12 +27,14 @@ import {Context} from 'src/clockface'
 
 // Actions
 import {deleteDemoDataBucketMembership} from 'src/cloud/actions/demodata'
+import {notify as notifyAction} from 'src/shared/actions/notifications'
 
 // Types
 import {DemoBucket} from 'src/types'
 
 interface DispatchProps {
   removeBucket: typeof deleteDemoDataBucketMembership
+  notify: typeof notifyAction
 }
 
 interface Props {
@@ -37,9 +46,24 @@ const DemoDataBucketCard: FC<Props & WithRouterProps & DispatchProps> = ({
   router,
   params: {orgID},
   removeBucket,
+  notify,
 }) => {
   const handleNameClick = () => {
     router.push(`/orgs/${orgID}/data-explorer?bucket=${bucket.name}`)
+  }
+
+  const handleCopyAttempt = (
+    copiedText: string,
+    isSuccessful: boolean
+  ): void => {
+    const text = copiedText.slice(0, 30).trimRight()
+    const truncatedText = `${text}...`
+
+    if (isSuccessful) {
+      notify(copyToClipboardSuccess(truncatedText, 'Bucket ID'))
+    } else {
+      notify(copyToClipboardFailed(truncatedText, 'Bucket ID'))
+    }
   }
 
   return (
@@ -83,6 +107,12 @@ const DemoDataBucketCard: FC<Props & WithRouterProps & DispatchProps> = ({
           Demo Data Bucket
         </span>
         <>Retention: {bucket.readableRetention}</>
+        <CopyToClipboard text={bucket.id} onCopy={handleCopyAttempt}>
+          <span className="copy-bucket-id" title="Click to Copy to Clipboard">
+            ID: {bucket.id}
+            <span className="copy-bucket-id--helper">Copy to Clipboard</span>
+          </span>
+        </CopyToClipboard>
       </ResourceCard.Meta>
       <FlexBox
         direction={FlexDirection.Row}
@@ -105,6 +135,7 @@ const DemoDataBucketCard: FC<Props & WithRouterProps & DispatchProps> = ({
 
 const mdtp: DispatchProps = {
   removeBucket: deleteDemoDataBucketMembership,
+  notify: notifyAction,
 }
 
 export default connect<{}, DispatchProps, {}>(
