+++ ui/src/style/chronograf.scss
@@ -123,6 +123,7 @@
 @import 'src/dashboards/components/DashboardsCardGrid.scss';
 @import 'src/dashboards/components/DashboardLightMode.scss';
 @import 'src/buckets/components/DemoDataDropdown.scss';
+@import 'src/buckets/components/BucketCardMeta.scss';
 @import 'src/shared/components/notifications/Notification.scss';
 
 // External
